pyproject.toml
[tool.poetry]
name = "backend-poc"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.1"
loguru = "0.7.2"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
python-dotenv = "^1.0.1"
dependency-injector = "^4.41.0"
pydantic-settings = "^2.2.1"
psycopg2 = "2.9.9"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
flake8 = "^7.0.0"
isort = "^5.13.2"

[tool.poetry.scripts]
backend = "python -m uvicorn backend_poc/main:app --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
